moved angleAverage to drakeUtil
